# Improve OptimizedSolver, add hole-checking to next rows