before: #3 

after:
```
test bench_3x20_optimized           ... bench:     173,366 ns/iter (+/- 6,275)
test bench_3x20_optimized_unique    ... bench:     174,413 ns/iter (+/- 1,875)
test bench_4x15_optimized           ... bench:   3,832,441 ns/iter (+/- 87,397)
test bench_4x15_optimized_unique    ... bench:   3,778,514 ns/iter (+/- 15,767)
test bench_5x12_optimized           ... bench:  20,870,829 ns/iter (+/- 2,242,749)
test bench_5x12_optimized_unique    ... bench:  20,367,612 ns/iter (+/- 3,486,165)
test bench_6x10_optimized           ... bench:  52,428,358 ns/iter (+/- 1,345,515)
test bench_6x10_optimized_unique    ... bench:  50,954,129 ns/iter (+/- 665,904)
test bench_8x8_2x2_optimized        ... bench:  11,657,095 ns/iter (+/- 126,585)
test bench_8x8_2x2_optimized_unique ... bench:  11,554,141 ns/iter (+/- 903,395)
```